Pre-Release-Version
Updated how authentication is handled. Updated readme and documentation in code.
