+++ bundestag_api/__init__.py
@@ -1,12 +1,7 @@
 # -*- coding: utf-8 -*-
 """
 Python wrapper for the official Bundestag API
-
-@author: Julian Schibberges
 """
 
-from .bta_utils import btapi_query, BTA_Connection, Person, Role, Drucksache, \
+from .bta_wrapper import BTA_Connection, Person, Role, Drucksache, \
     Aktivitaet, Vorgang, Vorgangsposition, Plenarprotokoll
-from .functions import search_procedure, search_document, search_person, \
-    search_plenaryprotocol, get_procedure, get_document, get_person, \
-    get_plenaryprotocol
